requirements.txt
Django==4.0.3
psycopg2-binary==2.9.3
python-decouple==3.6
djangorestframework==3.13.1
Pillow==9.1.0
